+++ myems-admin/app/services/users/user/user.service.js
@@ -112,6 +112,46 @@ app.factory('UserService', function($http) {
             }, function (response) {
                 callback(response);
             });
-        }
+        },
+        deleteNewUser: function(user, headers, callback) {  
+            $http.delete(getAPI()+'users/newusers/'+user.id, {headers})  
+            .then(function (response) {
+                callback(response);
+            }, function (response) {
+                callback(response);
+            });
+        },
+        editNewUser: function(user, headers, callback) {  
+            $http.put(getAPI()+'users/newusers/'+user.id, {"data":user}, {headers})  
+            .then(function (response) {
+                callback(response);
+            }, function (response) {
+                callback(response);
+            });
+        },
+        getAllNewUsers: function(headers, callback) {  
+            $http.get(getAPI()+'users/newusers', {headers})  
+            .then(function (response) {
+                callback(response);
+            }, function (response) {
+                callback(response);
+            }); 
+        },
+        getNewUser: function(id, headers, callback) {  
+            $http.get(getAPI()+'users/newusers/'+id, {headers})  
+            .then(function (response) {
+                callback(response);
+            }, function (response) {
+                callback(response);
+            }); 
+        },
+        approveUser: function(user, headers, callback){
+            $http.put(getAPI()+'users/newusers/'+user.id+'/approve', {"data":user}, {headers})
+            .then(function (response) {
+                callback(response);
+            }, function (response) {
+                callback(response);
+            });
+        },
     };
 });